fix: uso do context na base da festa e nos opcionais
